Simplify Crop and SelectToken implementations and tests
