+++ src/finn/custom_op/fpgadataflow/rtl/selecttoken_rtl.py
@@ -8,9 +8,6 @@ from finn.custom_op.fpgadataflow.selecttoken import SelectToken
 class SelectToken_rtl(SelectToken, Crop_rtl):
     """RTL SelectToken implemented by the shared Crop core."""
 
-    def __init__(self, onnx_node, **kwargs):
-        super().__init__(onnx_node, **kwargs)
-
     def get_nodeattr_types(self):
         return SelectToken.get_nodeattr_types(self) | Crop_rtl.get_nodeattr_types(self)
